# openapi.yaml
openapi: 3.1.0
info:
  title: NeoNet Client API
  version: 0.1.0
  description: |
    Local REST + WebSocket API exposed by the `neonet-node` daemon on
    `127.0.0.1:7780` (default).

    The daemon decrypts all payloads — client apps receive clear JSON and
    never handle cryptographic keys directly.
  license:
    name: AGPL-3.0-or-later
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: http://127.0.0.1:7780
    description: Local daemon (default port)
security:
  - BearerToken: []
...
paths:
...
  /v1/identity:
    get:
      operationId: getIdentity
      tags: [Identity]
      summary: Get local node identity.
      responses:
        "200":
          description: Node identity.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdentityResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/identity/sign:
    post:
      operationId: signPayload
      tags: [Identity]
      summary: Sign an arbitrary payload with the local private key.
      description: The private key never leaves the daemon.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignRequest"
      responses:
        "200":
          description: Signature produced.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ── Rooms ───────────────────────────────────────────────────────────
  /v1/rooms:
    get:
      operationId: listRooms
      tags: [Rooms]
      summary: List all rooms the node is a member of.
      responses:
        "200":
          description: Room list.
          content:
            application/json:
              schema:
                type: object
                required: [rooms]
                properties:
                  rooms:
                    type: array
                    items:
                      $ref: "#/components/schemas/RoomSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      operationId: createRoom
      tags: [Rooms]
      summary: Create a new room with initial members.
      description: |
        For `"direct"`: exactly 1 member in the list (+ self = 2).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoomRequest"
      responses:
        "200":
          description: Room created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateRoomResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/rooms/{room_id}:
    parameters:
      - $ref: "#/components/parameters/RoomId"

    get:
      operationId: getRoom
      tags: [Rooms]
      summary: Get full room state.
      responses:
        "200":
          description: Room detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      operationId: patchRoom
      tags: [Rooms]
      summary: Update room metadata (Admin/Owner only).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchRoomRequest"
      responses:
        "200":
          description: Room updated.
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      operationId: deleteRoom
      tags: [Rooms]
      summary: Leave a room. Owner deletes for everyone.
      responses:
        "200":
          description: Left / deleted.
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Messages ────────────────────────────────────────────────────────
  /v1/rooms/{room_id}/messages:
    parameters:
      - $ref: "#/components/parameters/RoomId"

    get:
      operationId: listMessages
      tags: [Messages]
      summary: List messages with cursor-based pagination.
      description: The daemon decrypts all payloads before returning.
      parameters:
        - name: since
          in: query
          description: Events after this event ID (exclusive).
          schema:
            type: string
        - name: before
          in: query
          description: Events before this event ID (exclusive).
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of messages (max 200).
          schema:
            type: integer
            default: 50
            maximum: 200
      responses:
        "200":
          description: Message list.
          content:
            application/json:
              schema:
                type: object
                required: [messages, has_more]
                properties:
                  messages:
                    type: array
                    items:
                      $ref: "#/components/schemas/DagEvent"
                  has_more:
                    type: boolean
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    post:
      operationId: sendMessage
      tags: [Messages]
      summary: Send a message to a room.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "200":
          description: Message created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendMessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/rooms/{room_id}/messages/{event_id}/react:
    parameters:
      - $ref: "#/components/parameters/RoomId"
      - $ref: "#/components/parameters/EventId"
    post:
      operationId: react
      tags: [Messages]
      summary: Add a reaction to a message.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReactRequest"
      responses:
        "200":
          description: Reaction created.
          content:
            application/json:
              schema:
                type: object
                required: [event_id]
                properties:
                  event_id:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/rooms/{room_id}/messages/{event_id}:
    parameters:
      - $ref: "#/components/parameters/RoomId"
      - $ref: "#/components/parameters/EventId"

    patch:
      operationId: editMessage
      tags: [Messages]
      summary: Edit a message (author only).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditMessageRequest"
      responses:
        "200":
          description: Edit event created.
          content:
            application/json:
              schema:
                type: object
                required: [event_id]
                properties:
                  event_id:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      operationId: deleteMessage
      tags: [Messages]
      summary: Delete a message (author or Admin/Owner).
      responses:
        "200":
          description: Redact event created.
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Members ─────────────────────────────────────────────────────────
...
  /v1/ws:
    get:
      operationId: websocket
      tags: [WebSocket]
      summary: Real-time event stream via WebSocket.
      description: |
        Upgrade to WebSocket with `Authorization: Bearer <token>`.

        One global WebSocket per app. After connecting, send `subscribe`
        messages to select which rooms to follow. The daemon pushes events
        for subscribed rooms only.

        **Client messages:** `WsSubscribe`, `WsUnsubscribe`
        **Server messages:** `WsNewMessage`, `WsMessageEdited`,
        `WsMessageDeleted`, `WsReaction`, `WsPresenceUpdate`,
        `WsMemberJoined`, `WsMemberLeft`, `WsRoomUpdated`, `WsSyncStatus`
      responses:
        "101":
          description: WebSocket upgrade successful.
        "401":
          $ref: "#/components/responses/Unauthorized"

# ─────────────────────────────────────────────────────────────────────
# Components
# ─────────────────────────────────────────────────────────────────────
components:
...
  securitySchemes:
    BearerToken:
      type: http
      scheme: bearer
      description: |
        Session token generated at daemon startup and written to
        `~/.neonet/session.token` (chmod 600). Rotated on every restart.

  # ── Parameters ──────────────────────────────────────────────────────
  parameters:
    RoomId:
      name: room_id
      in: path
      required: true
      description: BASE64URL-encoded room ID (blake3 hash).
      schema:
        type: string
    EventId:
      name: event_id
      in: path
      required: true
      description: BASE64URL-encoded event ID (blake3 hash).
      schema:
        type: string

  # ── Error responses ─────────────────────────────────────────────────
  responses:
    BadRequest:
      description: "400 — Missing or invalid parameters."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            error: bad_request
            message: Missing or invalid parameters
    Unauthorized:
      description: "401 — Missing or invalid session token."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            error: unauthorized
            message: Missing or invalid session token
    Forbidden:
      description: "403 — ACL violation."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            error: forbidden
            message: Action not authorized
    NotFound:
      description: "404 — Resource not found."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
          example:
            error: not_found
            message: Resource not found
...
  schemas:
...
    ApiError:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Machine-readable snake_case error code.
        message:
          type: string
          description: Human-readable description.
    NodeType:
      type: string
      enum: [full, relay, light]
    RoomType:
      type: string
      enum: [direct, group, channel, thread]
    Role:
      type: string
      enum: [owner, admin, member, guest]
    EventKind:
      type: string
      enum: [message, reaction, thread_reply, edit, redact, file, genesis]
    PresenceStatus:
      type: string
      enum: [online, away, offline]
...
    IdentityResponse:
      type: object
      required: [pubkey, address, node_type]
      properties:
        pubkey:
          type: string
          description: "`ed25519:BASE64URL` encoded public key."
          example: "ed25519:ABC123..."
        address:
          type: string
          description: "NeoNet address `@pubkey:domain`."
          example: "@ABC123...:mondomaine.com"
        node_type:
          $ref: "#/components/schemas/NodeType"
    SignRequest:
      type: object
      required: [payload]
      properties:
        payload:
          type: string
          description: BASE64URL-encoded bytes to sign.
    SignResponse:
      type: object
      required: [sig, pubkey]
      properties:
        sig:
          type: string
          description: "`ed25519:BASE64URL` encoded signature."
        pubkey:
          type: string
          description: "`ed25519:BASE64URL` encoded public key."

    # ── Rooms ───────────────────────────────────────────────────────
    RoomSummary:
      type: object
      required: [room_id, room_type, name]
      properties:
        room_id:
          type: string
        room_type:
          $ref: "#/components/schemas/RoomType"
        name:
          type: string
        description:
          type: string
        member_count:
          type: integer
        unread_count:
          type: integer
        last_event_id:
          type: string
        last_event_ts:
          type: integer
          description: Unix timestamp (seconds).
    CreateRoomRequest:
      type: object
      required: [name, room_type]
      properties:
        name:
          type: string
        description:
          type: string
        room_type:
          $ref: "#/components/schemas/RoomType"
        members:
          type: array
          items:
            type: string
          description: "NeoNet addresses of initial members."
    CreateRoomResponse:
      type: object
      required: [room_id, room_type, name, created_at]
      properties:
        room_id:
          type: string
        room_type:
          $ref: "#/components/schemas/RoomType"
        name:
          type: string
        created_at:
          type: integer
    RoomDetail:
      type: object
      required: [room_id, room_type, name, members, settings, created_at]
      properties:
        room_id:
          type: string
        room_type:
          $ref: "#/components/schemas/RoomType"
        name:
          type: string
        description:
          type: string
        members:
          type: array
          items:
            $ref: "#/components/schemas/MemberInfo"
        settings:
          type: object
          additionalProperties: true
        created_at:
          type: integer
    PatchRoomRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
    MemberInfo:
      type: object
      required: [address, role, joined_at]
      properties:
        address:
          type: string
          description: "NeoNet address `@pubkey:domain`."
        role:
          $ref: "#/components/schemas/Role"
        joined_at:
          type: integer
        presence:
          $ref: "#/components/schemas/PresenceStatus"

    # ── Messages ────────────────────────────────────────────────────
    DagEvent:
      type: object
      required: [event_id, kind, content, parents, ts_hint, edited, redacted]
      properties:
        event_id:
          type: string
        author:
          type: string
          description: "NeoNet address. Null in anonymous mode."
        kind:
          $ref: "#/components/schemas/EventKind"
        content:
          type: object
          additionalProperties: true
          description: Decrypted event payload (type depends on `kind`).
        parents:
          type: array
          items:
            type: string
          description: Causal parent event IDs (DAG edges).
        ts_hint:
          type: integer
          description: Non-authoritative client-declared timestamp.
        edited:
          type: boolean
        redacted:
          type: boolean
    SendMessageRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
        reply_to:
          type: string
          description: Optional parent event_id for thread replies.
    SendMessageResponse:
      type: object
      required: [event_id, ts_hint]
      properties:
        event_id:
          type: string
        ts_hint:
          type: integer
    ReactRequest:
      type: object
      required: [emoji]
      properties:
        emoji:
          type: string
    EditMessageRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string

    # ── Members ─────────────────────────────────────────────────────
...